+++ bpm-engine-impl/src/main/java/io/takari/bpm/elements/ServiceTaskHandler.java
@@ -9,14 +9,18 @@ import io.takari.bpm.actions.PopCommandAction;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.commands.Command;
 import io.takari.bpm.commands.ProcessElementCommand;
-import io.takari.bpm.model.*;
+import io.takari.bpm.model.ExpressionType;
+import io.takari.bpm.model.SequenceFlow;
+import io.takari.bpm.model.ServiceTask;
+import io.takari.bpm.model.VariableMapping;
 import io.takari.bpm.state.ProcessInstance;
 import io.takari.bpm.utils.Timeout;
-import org.joda.time.Duration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class ServiceTaskHandler implements ElementHandler {
 
@@ -39,14 +43,14 @@ public class ServiceTaskHandler implements ElementHandler {
             Command defaultCommand = new ProcessElementCommand(pd.getId(), nextFlow.getId());
 
             // collect all timeout boundary events attached to this task
-            List<Timeout<Command>> timeouts = findTimers(pd, cmd);
+            List<Timeout<Command>> timeouts = ProcessDefinitionUtils.findTimers(pd, cmd);
 
             // create a command which will be executed after an unspecified
             // error
-            Command defaultError = findDefaultError(pd, cmd);
+            Command defaultError = ProcessDefinitionUtils.findDefaultError(pd, cmd);
 
             // collect all boundary error events that have an errorRef
-            Map<String, Command> errors = findErrors(pd, cmd);
+            Map<String, Command> errors = ProcessDefinitionUtils.findErrors(pd, cmd);
 
             actions.add(new EvalExpressionAction.Builder(cmd.getDefinitionId(), t.getId(), t.getType(), t.getExpression(), defaultCommand)
                     .withTimeouts(timeouts)
@@ -70,45 +74,5 @@ public class ServiceTaskHandler implements ElementHandler {
             return null;
         }
         return s;
-    }
-
-    private static List<Timeout<Command>> findTimers(IndexedProcessDefinition pd, ProcessElementCommand cmd) throws ExecutionException {
-        List<BoundaryEvent> events = ProcessDefinitionUtils.findOptionalBoundaryEvents(pd, cmd.getElementId());
-        List<Timeout<Command>> l = new ArrayList<>(events.size());
-        for (BoundaryEvent ev : events) {
-            if (ev.getTimeDuration() != null) {
-                Duration d = Duration.parse(ev.getTimeDuration());
-                Command c = new ProcessElementCommand(pd.getId(), ev.getId());
-                l.add(new Timeout<>(d.getMillis(), c));
-            }
-        }
-
-        l.sort((o1, o2) -> (int) (o1.getDuration() - o2.getDuration()));
-
-        return l;
-    }
-
-    private static Command findDefaultError(IndexedProcessDefinition pd, ProcessElementCommand cmd) throws ExecutionException {
-        List<BoundaryEvent> events = ProcessDefinitionUtils.findOptionalBoundaryEvents(pd, cmd.getElementId());
-        for (BoundaryEvent ev : events) {
-            if (ev.getErrorRef() == null && ev.getTimeDuration() == null) {
-                return new ProcessElementCommand(pd.getId(), ev.getId());
-            }
-        }
-
-        return null;
-    }
-
-    private static Map<String, Command> findErrors(IndexedProcessDefinition pd, ProcessElementCommand cmd) throws ExecutionException {
-        Map<String, Command> m = new HashMap<>();
-
-        List<BoundaryEvent> events = ProcessDefinitionUtils.findOptionalBoundaryEvents(pd, cmd.getElementId());
-        for (BoundaryEvent ev : events) {
-            if (ev.getErrorRef() != null && ev.getTimeDuration() == null) {
-                m.put(ev.getErrorRef(), new ProcessElementCommand(pd.getId(), ev.getId()));
-            }
-        }
-
-        return m;
     }
 }